# docs/swagger.yaml
basePath: /
definitions:
  controller.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
      trace_id:
        type: object
    type: object
...
  model.DictType:
    properties:
      createBy:
        description: 创建者
        type: string
      createdAt:
        type: string
      dataScope:
        type: string
      deletedAt:
        type: string
      dictId:
        type: integer
      dictName:
        description: 字典名称
        type: string
      dictType:
        description: 字典类型
        type: string
      params:
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态
        type: string
      updateBy:
        description: 更新者
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Menu:
    properties:
      action:
        type: string
      breadcrumb:
        type: string
      children:
        items:
          $ref: '#/definitions/model.Menu'
        type: array
      component:
        type: string
      createBy:
        type: string
      createdAt:
        type: string
      dataScope:
        type: string
      deletedAt:
        type: string
      icon:
        type: string
      is_select:
        type: boolean
      isFrame:
        type: string
      menuId:
        type: integer
      menuName:
        type: string
      menuType:
        type: string
      noCache:
        type: boolean
      params:
        type: string
      parentId:
        type: integer
      path:
        type: string
      paths:
        type: string
      permission:
        type: string
      roleId:
        type: integer
      sort:
        type: integer
      title:
        type: string
      updateBy:
        type: string
      updatedAt:
        type: string
      visible:
        type: string
    type: object
...
host: 127.0.0.1:8060
info:
  contact:
    email: 1@lg1024.com
    name: lg1024
    url: http://lg1024.com
  description: 后台接口
  license: {}
  title: api-gin-web API
  version: 1.0.5
paths:
...
  /api/v1/db/columns/page:
    get:
      description: 数据库表列分页列表 / database table column page list
      parameters:
      - description: tableName / 数据表名称
        in: query
        name: tableName
        type: string
      - description: pageSize / 页条数
        in: query
        name: pageSize
        type: integer
      - description: pageIndex / 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 分页列表数据 / page list data
      tags:
      - 工具 / system
  /api/v1/db/tables/page:
    get:
      description: 数据库表分页列表 / database table page list
      parameters:
      - description: tableName / 数据表名称
        in: query
        name: tableName
        type: string
      - description: pageSize / 页条数
        in: query
        name: pageSize
        type: integer
      - description: pageIndex / 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 分页列表数据 / page list data
      tags:
      - 工具 / system
  /api/v1/dict/data:
    post:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DictType'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 添加字典数据
      tags:
      - 字典数据
    put:
      consumes:
      - application/json
      description: 获取JSON
      parameters:
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.DictType'
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 修改字典数据
      tags:
      - 字典数据
  /api/v1/dict/data/{dictCode}:
    delete:
      description: 删除数据
      parameters:
      - description: dictCode
        in: path
        name: dictCode
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "删除失败"}'
          schema:
            type: string
      summary: 删除字典数据
      tags:
      - 字典数据
    get:
      description: 获取JSON
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - "": []
      summary: 通过编码获取字典数据
      tags:
      - 字典数据
  /api/v1/dict/data/list:
    get:
      description: 获取JSON
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: dictCode
        in: query
        name: dictCode
        type: string
      - description: dictType
        in: query
        name: dictType
        type: string
      - description: 页条数
        in: query
        name: pageSize
        type: integer
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - "": []
      summary: 字典数据列表
      tags:
      - 字典数据
  /api/v1/dict/databyType/{dictType}:
    get:
      description: 获取JSON
      parameters:
      - description: dictType
        in: path
        name: dictType
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            $ref: '#/definitions/controller.Response'
      security:
      - "": []
      summary: 通过字典类型获取字典数据
      tags:
      - 字典数据
  /api/v1/menu:
    get:
      description: 获取JSON
      parameters:
      - description: menuName
        in: query
        name: menuName
        type: string
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: Menu列表数据
      tags:
      - 菜单
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取JSON
      parameters:
      - description: menuName
        in: formData
        name: menuName
        required: true
        type: string
      - description: Path
        in: formData
        name: Path
        type: string
      - description: Action
        in: formData
        name: Action
        required: true
        type: string
      - description: Permission
        in: formData
        name: Permission
        required: true
        type: string
      - description: ParentId
        in: formData
        name: ParentId
        required: true
        type: string
      - description: IsDel
        in: formData
        name: IsDel
        required: true
        type: string
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 创建菜单
      tags:
      - 菜单
  /api/v1/menu/{id}:
    delete:
      description: 删除数据
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "删除失败"}'
          schema:
            type: string
      summary: 删除菜单
      tags:
      - 菜单
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取JSON
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Menu'
      responses:
        "200":
          description: '{"code": -1, "message": "修改失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 修改菜单
      tags:
      - 菜单
  /api/v1/menuTreeselect:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取JSON
      responses:
        "200":
          description: '{"code": -1, "message": "添加失败"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 获取菜单树
      tags:
      - 菜单
  /api/v1/menuids/{id}:
    get:
      description: 获取JSON
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 获取角色对应的菜单id数组
      tags:
      - 菜单
  /api/v1/menulist:
    get:
      description: 获取JSON
      parameters:
      - description: menuName
        in: query
        name: menuName
        type: string
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: Menu列表数据
      tags:
      - 菜单
  /api/v1/menurole:
    get:
      description: 获取JSON
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code": -1, "message": "抱歉未找到相关信息"}'
          schema:
            type: string
      security:
      - Bearer: []
      summary: 根据角色名称获取菜单列表数据（左菜单使用）
      tags:
      - 菜单
...
schemes:
- http
- https
securityDefinitions:
  HeaderAuthorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"